requirements.txt
pandas==2.2.2
mysql-connector-python==8.3.0
pytest==8.1.1
